requirements.txt
matplotlib
pydicom
pyyaml
numpy
